+++ src/main/java/com/dsf/escalade/repository/business/VoieRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VoieRepository extends JpaRepository <Voie, Long> {
+public interface VoieRepository extends JpaRepository <Voie, Integer> {
 }
